requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.7.1
attrs==19.3.0
Click==7.0
coverage==5.0
cryptography==2.1.4
cycler==0.10.0
Flask==1.0.3
gast==0.2.2
gevent==1.4.0
google-pasta==0.1.7
greenlet==0.4.15
grpcio==1.20.1
gunicorn==19.9.0
h5py==2.9.0
idna==2.6
importlib-metadata==1.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
kafka==1.3.5
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
llvmlite==0.30.0
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==2.2.5
mock==3.0.5
more-itertools==5.0.0
natsort==6.0.0
numba==0.46.0
numpy==1.16.0
opencv-python==3.4.1.15
ordereddict==1.1
packaging==19.2
pluggy==0.13.1
ply==3.11
protobuf==3.7.1
py==1.8.0
py-cpuinfo==5.0.0
py-eureka-client==0.5.7
pycrypto==2.6.1
pyparsing==2.4.0
pytest==4.6.11
pytest-benchmark==3.2.2
python-dateutil==2.8.0
pyxdg==0.25
PyYAML==3.13
requests==2.9.1
scikit-learn==0.20.3
scipy==1.2.0
SecretStorage==2.3.1
Shapely==1.6.4.post2
six==1.11.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
thriftpy2==0.4.8
torch==1.2.0
torchvision==0.4.0
tqdm==4.40.2
wcwidth==0.1.7
Werkzeug==0.15.4
wrapt==1.11.2
zipp==0.6.0
